+++ storybook/src/components/FieldSet/FieldSet.stories.tsx
@@ -22,6 +22,8 @@ const meta = {
   args: {
     invalid: false,
     legend: 'Wat is uw naam?',
+    optional: false,
+    hint: '',
   },
   decorators: [
     (Story) => (
@@ -37,26 +39,26 @@ export default meta
 type Story = StoryObj<typeof meta>
 
 export const Default: Story = {
-  render: (args) => (
-    <FieldSet invalid={args.invalid} legend={args.legend}>
+  render: ({ invalid, legend, hint, optional }) => (
+    <FieldSet hint={hint} invalid={invalid} legend={legend} optional={optional}>
       <Field className="ams-mb--sm">
         <Label htmlFor="input-a1">Voornaam</Label>
-        {args.invalid && <ErrorMessage id="error-a1">Vul uw voornaam in.</ErrorMessage>}
+        {invalid && <ErrorMessage id="error-a1">Vul uw voornaam in.</ErrorMessage>}
         <TextInput
-          aria-describedby={args.invalid ? 'error-a1' : undefined}
+          aria-describedby={invalid ? 'error-a1' : undefined}
           aria-required="true"
           id="input-a1"
-          invalid={args.invalid}
+          invalid={invalid}
         />
       </Field>
       <Field>
         <Label htmlFor="input-a2">Achternaam</Label>
-        {args.invalid && <ErrorMessage id="error-a2">Vul uw achternaam in.</ErrorMessage>}
+        {invalid && <ErrorMessage id="error-a2">Vul uw achternaam in.</ErrorMessage>}
         <TextInput
-          aria-describedby={args.invalid ? 'error-a2' : undefined}
+          aria-describedby={invalid ? 'error-a2' : undefined}
           aria-required="true"
           id="input-a2"
-          invalid={args.invalid}
+          invalid={invalid}
         />
       </Field>
     </FieldSet>
@@ -65,7 +67,7 @@ export const Default: Story = {
 
 export const WithDescription: Story = {
   render: (args) => (
-    <FieldSet invalid={args.invalid} legend={args.legend} aria-describedby="description-b">
+    <FieldSet aria-describedby="description-b" invalid={args.invalid} legend={args.legend}>
       <Paragraph id="description-b" size="small" className="ams-mb--sm">
         Vul uw naam in zoals in uw paspoort staat.
       </Paragraph>
@@ -93,10 +95,38 @@ export const WithDescription: Story = {
   ),
 }
 
-export const WithError: Story = {
+export const WithHint: Story = {
+  args: { optional: false, hint: 'verplicht' },
+  render: ({ invalid, legend, hint, optional }) => (
+    <FieldSet hint={hint} invalid={invalid} legend={legend} optional={optional}>
+      <Field className="ams-mb--sm">
+        <Label htmlFor="input-b3">Voornaam</Label>
+        {invalid && <ErrorMessage id="error-b3">Vul uw voornaam in.</ErrorMessage>}
+        <TextInput
+          aria-describedby={invalid ? 'error-b3' : undefined}
+          aria-required="true"
+          id="input-b3"
+          invalid={invalid}
+        />
+      </Field>
+      <Field>
+        <Label htmlFor="input-b4">Achternaam</Label>
+        {invalid && <ErrorMessage id="error-b4">Vul uw achternaam in.</ErrorMessage>}
+        <TextInput
+          aria-describedby={invalid ? 'error-b4' : undefined}
+          aria-required="true"
+          id="input-b4"
+          invalid={invalid}
+        />
+      </Field>
+    </FieldSet>
+  ),
+}
+
+export const WithValidation: Story = {
   args: { invalid: true },
   render: (args) => (
-    <FieldSet invalid={args.invalid} legend={args.legend} aria-describedby="description-c">
+    <FieldSet aria-describedby="description-c" invalid={args.invalid} legend={args.legend}>
       <Paragraph id="description-c" size="small" className="ams-mb--sm">
         Vul uw naam in zoals in uw paspoort staat.
       </Paragraph>
@@ -162,7 +192,7 @@ export const RadioGroup: Story = {
   ),
 }
 
-export const RadioGroupWithError: Story = {
+export const RadioGroupWithValidation: Story = {
   args: {
     legend: 'Waar gaat uw melding over?',
     invalid: true,
@@ -238,7 +268,7 @@ export const CheckboxGroup: Story = {
   ),
 }
 
-export const CheckboxGroupWithError: Story = {
+export const CheckboxGroupWithValidation: Story = {
   args: {
     invalid: true,
     legend: 'Waar gaat uw melding over?',