+++ storybook/src/components/Label/Label.stories.tsx
@@ -11,6 +11,8 @@ const meta = {
   component: Label,
   args: {
     children: 'Label',
+    hint: '',
+    optional: false,
   },
   argTypes: {
     children: {
@@ -25,3 +27,15 @@ export default meta
 type Story = StoryObj<typeof meta>
 
 export const Default: Story = {}
+
+export const Optional: Story = {
+  args: {
+    optional: true,
+  },
+}
+
+export const WithHint: Story = {
+  args: {
+    hint: 'not required',
+  },
+}